requirements.txt
# ==========================
# Telegram Bot
# ==========================
python-telegram-bot>=20.7

# ==========================
# Web framework & environment
# ==========================
flask>=2.3.3
python-dotenv>=1.2.1
requests>=2.32.5
aiohttp>=3.9.0
python-multipart>=0.0.6

# ==========================
# Blockchain support
# ==========================
web3>=6.8.0
eth-utils>=2.3.0
eth-account>=0.9.0

# ==========================
# Database / ORM
# ==========================
SQLAlchemy>=2.0.36
